add support for Decimal config param, remove node instance ID as input, and move helper functions to node.py
