+++ .coveragerc
@@ -1032,6 +1032,7 @@ omit =
     homeassistant/components/zwave/util.py
     homeassistant/components/ozw/__init__.py
     homeassistant/components/ozw/entity.py
+    homeassistant/components/ozw/node.py
     homeassistant/components/ozw/services.py
 
 [report]
